fix: refactor jwks base url property name (#13)
