+++ helm/values-dev.yaml
@@ -31,6 +31,10 @@ microservice-chart:
     ENV: "DEV"
     JAVA_TOOL_OPTIONS: "-Xms128m -Xmx4g -Djava.util.concurrent.ForkJoinPool.common.parallelism=7 -javaagent:/app/applicationinsights-agent.jar -Dapplicationinsights.configuration.file=/mnt/file-config-external/appinsights-config/applicationinsights.json -agentlib:jdwp=transport=dt_socket,server=y,address=8001,suspend=n -Dcom.sun.management.jmxremote=true -Dcom.sun.management.jmxremote.port=3002 -Dcom.sun.management.jmxremote.rmi.port=3003 -Djava.rmi.server.hostname=127.0.0.1 -Dcom.sun.management.jmxremote.authenticate=false -Dcom.sun.management.jmxremote.ssl=false"
 
+    JWT_TOKEN_AUDIENCE: "dev.piattaformaunitaria.pagopa.it"
+    JWT_EXTERNAL_TOKEN_ISS: "https://dev.selfcare.pagopa.it"
+    JWT_EXTERNAL_TOKEN_BASE_URL: "https://dev.selfcare.pagopa.it"
+
   keyvault:
     name: "p4pa-d-payhub-kv"
     tenantId: "7788edaf-0346-4068-9d79-c868aed15b3d"
