+++ helm/values-uat.yaml
@@ -31,6 +31,9 @@ microservice-chart:
     ENV: "UAT"
     JAVA_TOOL_OPTIONS: "-Xms128m -Xmx4g -Djava.util.concurrent.ForkJoinPool.common.parallelism=7 -javaagent:/app/applicationinsights-agent.jar -Dapplicationinsights.configuration.file=/mnt/file-config-external/appinsights-config/applicationinsights.json -agentlib:jdwp=transport=dt_socket,server=y,address=8001,suspend=n -Dcom.sun.management.jmxremote=true -Dcom.sun.management.jmxremote.port=3002 -Dcom.sun.management.jmxremote.rmi.port=3003 -Djava.rmi.server.hostname=127.0.0.1 -Dcom.sun.management.jmxremote.authenticate=false -Dcom.sun.management.jmxremote.ssl=false"
 
+    JWT_TOKEN_AUDIENCE: "uat.piattaformaunitaria.pagopa.it"
+    JWT_EXTERNAL_TOKEN_ISS: "https://uat.selfcare.pagopa.it"
+    JWT_EXTERNAL_TOKEN_BASE_URL: "https://uat.selfcare.pagopa.it"
 
   keyvault:
     name: "p4pa-u-payhub-kv"
